+++ src/main/java/org/jenkinsci/plugins/DependencyTrack/api/ApiClient.java
@@ -255,17 +255,38 @@ public class ApiClient {
 
     @Nonnull
     public List<Violation> getViolations(@Nonnull final String projectUuid) throws ApiClientException {
-        final var uri = UriComponentsBuilder.fromUriString(PROJECT_VIOLATIONS_URL).pathSegment("{uuid}").build(projectUuid);
+        final List<Violation> violations = new ArrayList<>();
+        int page = 1;
+        boolean fetchMore = true;
+        while (fetchMore) {
+            var fetchedViolations = getViolationsPaged(projectUuid, page++);
+            violations.addAll(fetchedViolations.result());
+            // Continue to retrieve further violations if the current result was not empty and the total amount has not yet been reached.
+            fetchMore = !fetchedViolations.isEmpty() && violations.size() < fetchedViolations.totalSize();
+        }
+        return violations;
+    }
+
+    @Nonnull
+    @SuppressWarnings("unchecked")
+    private PagedResult<Violation> getViolationsPaged(@Nonnull final String projectUuid, final int page) throws ApiClientException {
+        final var uri = UriComponentsBuilder.fromUriString(PROJECT_VIOLATIONS_URL).pathSegment("{uuid}")
+                .queryParam(PAGINATED_REQ_PAGE_PARAM, "{page}")
+                .queryParam(PAGINATED_REQ_PAGESIZE_PARAM, 100)
+                .build(projectUuid, page);
         final var request = createRequest(uri);
         return executeWithRetry(() -> {
             try (var response = httpClient.newCall(request).execute()) {
                 final var body = response.body().string();
-                if (!response.isSuccessful()) {
+                if (response.isSuccessful()) {
+                    final var violations = ViolationParser.parse(body);
+                    final var totalCount = getTotalCountValue(response, violations.size());
+                    return new PagedResult<>(violations, totalCount);
+                } else {
                     final int status = response.code();
                     logger.log(body);
                     throw new ApiClientException(Messages.ApiClient_Error_RetrieveViolations(status, HttpStatus.valueOf(status).getReasonPhrase()));
                 }
-                return ViolationParser.parse(body);
             } catch (ApiClientException e) {
                 throw e;
             } catch (IOException e) {