+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ⭐ New Features
 ### 🐞 Bugs Fixed
 - The list of projects in the select box did not show all available projects.
+- Policy violations were limited to 100 ([#345](https://github.com/jenkinsci/dependency-track-plugin/issues/345))
 
 ## v6.0.1 - 2025-03-23
 ### ⚠ Breaking